Fix issue #23  - Remove custom channel types per channel  - Remove channel groups as they were not useful  - Improve support for dynamic channels  - Fix issues with channel linking in the UI  - Refactor some code around
