+++ src/main/java/org/openhab/binding/solarman/internal/updater/SolarmanChannelUpdater.java
@@ -1,0 +1,226 @@
+package org.openhab.binding.solarman.internal.updater;
+
+import org.eclipse.jdt.annotation.Nullable;
+import org.openhab.binding.solarman.internal.defmodel.ParameterItem;
+import org.openhab.binding.solarman.internal.defmodel.Request;
+import org.openhab.binding.solarman.internal.defmodel.Validation;
+import org.openhab.binding.solarman.internal.modbus.SolarmanLoggerConnection;
+import org.openhab.binding.solarman.internal.modbus.SolarmanLoggerConnector;
+import org.openhab.binding.solarman.internal.modbus.SolarmanV5Protocol;
+import org.openhab.binding.solarman.internal.util.StreamUtils;
+import org.openhab.core.library.types.DateTimeType;
+import org.openhab.core.library.types.DecimalType;
+import org.openhab.core.library.types.QuantityType;
+import org.openhab.core.library.types.StringType;
+import org.openhab.core.thing.ChannelUID;
+import org.openhab.core.thing.ThingStatus;
+import org.openhab.core.types.State;
+import org.openhab.core.types.Type;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import tech.units.indriya.format.SimpleUnitFormat;
+
+import javax.measure.Unit;
+import javax.measure.format.MeasurementParseException;
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
+
+import static org.openhab.binding.solarman.internal.util.StreamUtils.reverse;
+
+public class SolarmanChannelUpdater {
+    private final Logger logger = LoggerFactory.getLogger(SolarmanChannelUpdater.class);
+    private final StatusUpdater statusUpdater;
+    private final StateUpdater stateUpdater;
+
+    public SolarmanChannelUpdater(StatusUpdater statusUpdater, StateUpdater stateUpdater) {
+        this.statusUpdater = statusUpdater;
+        this.stateUpdater = stateUpdater;
+    }
+
+    public void fetchDataFromLogger(List<Request> requests,
+                                    SolarmanLoggerConnector solarmanLoggerConnector,
+                                    SolarmanV5Protocol solarmanV5Protocol,
+                                    Map<ParameterItem, ChannelUID> paramToChannelMapping) {
+
+        try (SolarmanLoggerConnection solarmanLoggerConnection = solarmanLoggerConnector.createConnection()) {
+            logger.debug("Fetching data from logger");
+
+            Map<Integer, byte[]> readRegistersMap = requests.stream()
+                    .map(request -> solarmanV5Protocol.readRegisters(solarmanLoggerConnection,
+                            (byte) request.getMbFunctioncode().intValue(),
+                            request.getStart(),
+                            request.getEnd())
+                    )
+                    .reduce(new HashMap<>(), this::mergeMaps);
+
+            updateChannelsForReadRegisters(paramToChannelMapping, readRegistersMap);
+
+            statusUpdater.updateStatus(readRegistersMap.isEmpty() ? ThingStatus.OFFLINE : ThingStatus.ONLINE);
+        } catch (Exception e) {
+            logger.error("Error invoking handler", e);
+        }
+    }
+
+    private void updateChannelsForReadRegisters(Map<ParameterItem, ChannelUID> paramToChannelMapping,
+                                                Map<Integer, byte[]> readRegistersMap) {
+        paramToChannelMapping.forEach((parameterItem, channelUID) -> {
+            List<Integer> registers = parameterItem.getRegisters();
+            if (readRegistersMap.keySet().containsAll(registers)) {
+                switch (parameterItem.getRule()) {
+                    case 1, 3 -> updateChannelWithNumericValue(parameterItem, channelUID, registers,
+                            readRegistersMap, ValueType.UNSIGNED);
+                    case 2, 4 -> updateChannelWithNumericValue(parameterItem, channelUID, registers,
+                            readRegistersMap, ValueType.SIGNED);
+                    case 5 -> updateChannelWithStringValue(channelUID, registers, readRegistersMap);
+                    case 6 -> updateChannelWithRawValue(parameterItem, channelUID, registers, readRegistersMap);
+                    case 7 -> updateChannelWithVersion(channelUID, registers, readRegistersMap);
+                    case 8 -> updateChannelWithDateTime(channelUID, registers, readRegistersMap);
+                    case 9 -> updateChannelWithTime(channelUID, registers, readRegistersMap);
+                }
+            } else {
+                logger.error("Unable to update channel {} because its registers were not read", channelUID.getId());
+            }
+        });
+    }
+
+    private void updateChannelWithTime(ChannelUID channelUID, List<Integer> registers, Map<Integer, byte[]> readRegistersMap) {
+        String stringValue = registers.stream()
+                .map(readRegistersMap::get)
+                .map(v -> ByteBuffer.wrap(v).getShort())
+                .map(rawVal -> String.format("%02d", rawVal / 100) + ":" +
+                        String.format("%02d", rawVal % 100))
+                .collect(Collectors.joining());
+
+        stateUpdater.updateState(channelUID, new StringType(stringValue));
+    }
+
+    private void updateChannelWithDateTime(ChannelUID channelUID, List<Integer> registers, Map<Integer, byte[]> readRegistersMap) {
+        String stringValue = StreamUtils.zip(
+                        IntStream.range(0, registers.size()).boxed(),
+                        registers.stream().map(readRegistersMap::get).map(v -> ByteBuffer.wrap(v).getShort()),
+                        StreamUtils.Tuple::new)
+                .map(t -> {
+                    int index = t.a();
+                    short rawVal = t.b();
+
+                    return switch (index) {
+                        case 0 -> (rawVal >> 8) + "/" + (rawVal & 0xFF) + "/";
+                        case 1 -> (rawVal >> 8) + " " + (rawVal & 0xFF) + ":";
+                        case 2 -> (rawVal >> 8) + ":" + (rawVal & 0xFF);
+                        default -> (rawVal >> 8) + "" + (rawVal & 0xFF);
+                    };
+                })
+                .collect(Collectors.joining());
+
+        try {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yy/M/d H:m:s");
+            LocalDateTime dateTime = LocalDateTime.parse(stringValue, formatter);
+
+            stateUpdater.updateState(channelUID, new DateTimeType(dateTime.atZone(ZoneId.systemDefault())));
+        } catch (DateTimeParseException e) {
+            logger.error("Unable to parse string date {} to a DateTime object", stringValue);
+        }
+    }
+
+    private void updateChannelWithVersion(ChannelUID channelUID, List<Integer> registers, Map<Integer, byte[]> readRegistersMap) {
+        String stringValue = registers.stream()
+                .map(readRegistersMap::get)
+                .map(v -> ByteBuffer.wrap(v).getShort())
+                .map(rawVal -> (rawVal >> 12) + "." +
+                        ((rawVal >> 8) & 0x0F) + "." +
+                        ((rawVal >> 4) & 0x0F) + "." +
+                        (rawVal & 0x0F))
+                .collect(Collectors.joining());
+
+        stateUpdater.updateState(channelUID, new StringType(stringValue));
+    }
+
+    private void updateChannelWithStringValue(ChannelUID channelUID, List<Integer> registers, Map<Integer, byte[]> readRegistersMap) {
+        String stringValue = registers.stream().map(readRegistersMap::get).reduce(new StringBuilder(), (acc, val) -> {
+            short shortValue = ByteBuffer.wrap(val).order(ByteOrder.BIG_ENDIAN).getShort();
+            return acc.append((char) (shortValue >> 8)).append((char) (shortValue & 0xFF));
+        }, StringBuilder::append).toString();
+
+        stateUpdater.updateState(channelUID, new StringType(stringValue));
+    }
+
+    private void updateChannelWithNumericValue(ParameterItem parameterItem, ChannelUID channelUID,
+                                               List<Integer> registers, Map<Integer, byte[]> readRegistersMap, ValueType valueType) {
+        BigInteger value = extractNumericValue(registers, readRegistersMap, valueType);
+        BigDecimal convertedValue = convertNumericValue(value, parameterItem.getOffset(), parameterItem.getScale());
+        if (validateNumericValue(convertedValue, parameterItem.getValidation())) {
+            State state;
+            if (parameterItem.getUom() != null) {
+                try {
+                    Unit<?> uom = SimpleUnitFormat.getInstance().parse(parameterItem.getUom());
+                    state = new QuantityType<>(convertedValue, uom);
+                } catch (MeasurementParseException e) {
+                    state = new DecimalType(convertedValue);
+                }
+
+            } else {
+                state = new DecimalType(convertedValue);
+            }
+            stateUpdater.updateState(channelUID, state);
+        }
+    }
+
+    private void updateChannelWithRawValue(ParameterItem parameterItem, ChannelUID channelUID, List<Integer> registers,
+                                           Map<Integer, byte[]> readRegistersMap) {
+        String hexString = String.format("[%s]",
+                reverse(registers).stream()
+                        .map(readRegistersMap::get)
+                        .map(val -> String.format("0x%02X", ByteBuffer.wrap(val).order(ByteOrder.BIG_ENDIAN).getShort()))
+                        .collect(Collectors.joining(",")));
+
+        stateUpdater.updateState(channelUID, new StringType(hexString));
+    }
+
+
+    private boolean validateNumericValue(BigDecimal convertedValue, Validation validation) {
+        return true;
+    }
+
+    private BigDecimal convertNumericValue(BigInteger value, @Nullable BigDecimal offset, @Nullable BigDecimal scale) {
+        return new BigDecimal(value).subtract(offset != null ? offset : BigDecimal.ZERO)
+                .multiply(scale != null ? scale : BigDecimal.ONE);
+    }
+
+    private BigInteger extractNumericValue(List<Integer> registers, Map<Integer, byte[]> readRegistersMap, ValueType valueType) {
+        return reverse(registers).stream().map(readRegistersMap::get).reduce(BigInteger.ZERO,
+                (acc, val) -> acc.shiftLeft(Short.SIZE).add(BigInteger.valueOf(ByteBuffer.wrap(val).getShort() & (valueType == ValueType.UNSIGNED ? 0xFFFF : 0xFFFFFFFF))),
+                BigInteger::add);
+    }
+
+    private <K, V> Map<K, V> mergeMaps(Map<K, V> map1,
+                                       Map<K, V> map2) {
+        return Stream.concat(map1.entrySet().stream(), map2.entrySet().stream())
+                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (v1, v2) -> v1));
+    }
+
+    private enum ValueType {
+        UNSIGNED, SIGNED
+    }
+
+    @FunctionalInterface
+    public interface StatusUpdater {
+        void updateStatus(ThingStatus thingStatus);
+    }
+
+    @FunctionalInterface
+    public interface StateUpdater {
+        void updateState(ChannelUID channelUID, State state);
+    }
+}
