+++ src/main/resources/OH-INF/config/datetime-channel-config.xml
@@ -4,7 +4,7 @@
 	xmlns:config-description="https://openhab.org/schemas/config-description/v1.0.0"
 	xsi:schemaLocation="https://openhab.org/schemas/config-description/v1.0.0 https://openhab.org/schemas/config-description-1.0.0.xsd">
 
-	<config-description uri="channel-type-config:solarman:string_channel">
+	<config-description uri="channel-type-config:solarman:datetime_channel">
 		<parameter name="uom" type="text">
 			<label>Unit of Measurement</label>
 			<description>The unit of measurement used for this channel</description>
