+++ src/main/resources/OH-INF/thing/channels.xml
@@ -7,13 +7,19 @@
 	<channel-type id="string">
 		<item-type>String</item-type>
 		<label>Text Value</label>
-		<config-description-ref uri="thing-type:solarman:string_channel"/>
+		<config-description-ref uri="channel-type-config:solarman:string_channel"/>
 	</channel-type>
 
 	<channel-type id="number">
 		<item-type>Number</item-type>
 		<label>Number Value</label>
-		<config-description-ref uri="thing-type:solarman:number_channel"/>
+		<config-description-ref uri="channel-type-config:solarman:number_channel"/>
+	</channel-type>
+
+	<channel-type id="datetime">
+		<item-type>Datetime</item-type>
+		<label>Datetime Value</label>
+		<config-description-ref uri="channel-type-config:solarman:datetime_channel"/>
 	</channel-type>
 
 </thing:thing-descriptions>
