+++ src/main/resources/OH-INF/thing/thing-types.xml
@@ -1,53 +1,68 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <thing:thing-descriptions bindingId="solarman"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:thing="https://openhab.org/schemas/thing-description/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/thing-description/v1.0.0 https://openhab.org/schemas/thing-description-1.0.0.xsd">
+                          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                          xmlns:thing="https://openhab.org/schemas/thing-description/v1.0.0"
+                          xsi:schemaLocation="https://openhab.org/schemas/thing-description/v1.0.0 https://openhab.org/schemas/thing-description-1.0.0.xsd">
 
-	<!-- Sample Thing Type -->
-	<thing-type id="logger" extensible="true">
-		<label>Solarman Binding Thing</label>
-		<description>Solarman Logger Binding Thing</description>
+    <!-- Sample Thing Type -->
+    <thing-type id="logger" extensible="true">
+        <label>Solarman Binding Thing</label>
+        <description>Solarman Logger Binding Thing</description>
 
-		<config-description>
-			<parameter name="hostname" type="text" required="true">
-				<context>network-address</context>
-				<label>Logger hostname</label>
-				<description>Hostname or IP address of the Solarman logger</description>
-				<advanced>false</advanced>
-			</parameter>
-			<parameter name="port" type="integer" required="false">
-				<label>Logger port</label>
-				<description>Port of the Solarman logger (default 8899)</description>
-				<default>8899</default>
-				<advanced>true</advanced>
-			</parameter>
-			<parameter name="inverterType" type="text" required="false">
-				<label>Inverter type</label>
-				<description>The type of inverter connected to the logger (default deye_sg04lp3)</description>
-				<default>deye_sg04lp3</default>
-				<advanced>false</advanced>
-			</parameter>
-			<parameter name="serialNumber" type="text" required="true">
-				<label>Logger serial number</label>
-				<description>Serial number of the Solarman logger</description>
-				<advanced>false</advanced>
-			</parameter>
-			<parameter name="refreshInterval" type="integer" required="false" unit="s" min="30">
-				<label>Refresh interval</label>
-				<description>Interval to query the logger (default 60)</description>
-				<default>60</default>
-				<advanced>true</advanced>
-			</parameter>
-			<parameter name="additionalRequests" type="text" required="false">
-				<label>Additional Requests</label>
-				<description>Additional requests besides the ones defined in the inverter definition.
-					Format is mb_functioncode1:start1-end1, mb_functioncode2:start2-end2,...
-					Example 0x03:0x0000-0x0100,0x03:0x0200-0x0300
-				</description>
-				<advanced>true</advanced>
-			</parameter>
-		</config-description>
-	</thing-type>
+        <config-description>
+            <parameter name="hostname" type="text" required="true">
+                <context>network-address</context>
+                <label>Logger hostname</label>
+                <description>Hostname or IP address of the Solarman logger</description>
+                <advanced>false</advanced>
+            </parameter>
+            <parameter name="port" type="integer" required="false">
+                <label>Logger port</label>
+                <description>Port of the Solarman logger (default 8899)</description>
+                <default>8899</default>
+                <advanced>true</advanced>
+            </parameter>
+            <parameter name="inverterType" type="text" required="false">
+                <label>Inverter type</label>
+                <description>The type of inverter connected to the logger (default deye_sg04lp3)</description>
+                <default>deye_sg04lp3</default>
+                <advanced>false</advanced>
+                <options>
+                    <option value="deye_2mppt">DEYE Microinverter with 2 MPPT Trackers (deye_2mppt)</option>
+                    <option value="deye_4mppt">DEYE Microinverter with 4 MPPT Trackers (deye_4mppt)</option>
+                    <option value="deye_hybrid">Generic DEYE/Sunsynk/SolArk Hybrid inverters (deye_hybrid)</option>
+                    <option value="deye_sg04lp3">DEYE/Sunsynk/SolArk Hybrid 8/12K-SG04LP3 (deye_sg04lp3)</option>
+                    <option value="deye_string">Generic DEYE/Sunsynk/SolArk String inverters (deye_string)</option>
+                    <option value="kstar_hybrid">KSTAR Hybrid Inverter (kstar_hybrid)</option>
+                    <option value="sofar_g3hyd">SOFAR Hybrid Three-Phase inverter (sofar_g3hyd)</option>
+                    <option value="sofar_hyd3k-6k-es">SOFAR Hybrid Single-Phase inverter (sofar_hyd3k-6k-es)</option>
+                    <option value="sofar_lsw3">SOFAR Inverters (sofar_lsw3)</option>
+                    <option value="sofar_wifikit">SOFAR WifiKit (sofar_wifikit)</option>
+                    <option value="solis_1p8k-5g">SOLIS 1P8K-5G (solis_1p8k-5g)</option>
+                    <option value="solis_hybrid">SOLIS Hybrid inverter (solis_hybrid)</option>
+                    <option value="zcs_azzurro-ktl-v3">ZCS Azzurro KTL-V3 inverters (zcs_azzurro-ktl-v3)</option>
+                </options>
+            </parameter>
+            <parameter name="serialNumber" type="text" required="true">
+                <label>Logger serial number</label>
+                <description>Serial number of the Solarman logger</description>
+                <advanced>false</advanced>
+            </parameter>
+            <parameter name="refreshInterval" type="integer" required="false" unit="s" min="30">
+                <label>Refresh interval</label>
+                <description>Interval to query the logger (default 60)</description>
+                <default>60</default>
+                <advanced>true</advanced>
+            </parameter>
+            <parameter name="additionalRequests" type="text" required="false">
+                <label>Additional Requests</label>
+                <description>Additional requests besides the ones defined in the inverter definition.
+                    Format is mb_functioncode1:start1-end1, mb_functioncode2:start2-end2,...
+                    Example 0x03:0x0000-0x0100,0x03:0x0200-0x0300
+                </description>
+                <advanced>true</advanced>
+            </parameter>
+        </config-description>
+    </thing-type>
 
 </thing:thing-descriptions>
